everythign seems to work, have pslit up everyhting except createQuiz into user server and use client , create is outside of game so doenst really need the perf boost atm.
